move handling of moose pp files into moose lib
from moose bin
